extracted nav-items into single file for single source of truth usage in mobile and desktop
