+++ src/app/page.tsx
@@ -6,7 +6,7 @@ export default function Home() {
 			<Link href={'/registration'}>Go to Registration</Link>
 			<Link href={'/login'}>Go to Login</Link>
 			<Link href={'/terminplan'}>Go to Terminplan</Link>
-			<Link href={'/vorlesungsuebersicht'}>Go to Vorlesungsübersicht</Link>
+			<Link href={'/vorlesungen'}>Go to Vorlesungen</Link>
 		</div>
 	);
 }
