+++ src/frontend/components/navigation/main-nav.tsx
@@ -6,44 +6,13 @@ import React from 'react';
 import Link from 'next/link';
 import { usePathname } from 'next/navigation';
 
-const navItems: { title: string; href: string; description: string }[] = [
-	{
-		title: 'Home',
-		href: '/',
-		description: 'Home page',
-	},
-	{
-		title: 'Registration',
-		href: '/registration',
-		description: 'Registration page',
-	},
-	{
-		title: 'Login',
-		href: '/login',
-		description: 'Login page',
-	},
-	{
-		title: 'Terminkalender',
-		href: '/terminplan',
-		description: 'Übersicht aller wichtigen Termine der Hochschule',
-	},
-	{
-		title: 'Vorlesungsübersicht',
-		href: '/vorlesungsuebersicht',
-		description: 'Übersicht aller Vorlesungen der ersten zwei Semester',
-	},
-	{
-		title: 'Not Found',
-		href: '/not-found',
-		description: 'Übersicht aller wichtigen Termine der Hochschule',
-	},
-];
+import navItems from '@/frontend/components/navigation/nav-items';
 
 const MainNav = () => {
 	const pathname = usePathname();
 
 	return (
-		<div className="hidden md:flex justify-between w-full max-w-[960px]">
+		<div className="hidden w-full max-w-[960px] justify-between md:flex">
 			<Link
 				href="/"
 				className="mr-6 flex items-center space-x-2"
@@ -57,7 +26,10 @@ const MainNav = () => {
 					<Link
 						key={item.href}
 						href={item.href}
-						className={cn('transition-colors hover:text-foreground/90', pathname === item.href ? 'text-foreground' : 'text-foreground/60')}
+						className={cn(
+							'transition-colors hover:text-foreground/90',
+							pathname === item.href ? 'text-foreground' : 'text-foreground/60'
+						)}
 					>
 						{item.title}
 					</Link>
