+++ src/frontend/components/navigation/nav-items.ts
@@ -1,0 +1,34 @@
+const navItems: { title: string; href: string; description: string }[] = [
+	{
+		title: 'Home',
+		href: '/',
+		description: 'Home page',
+	},
+	{
+		title: 'Registration',
+		href: '/registration',
+		description: 'Registration page',
+	},
+	{
+		title: 'Login',
+		href: '/login',
+		description: 'Login page',
+	},
+	{
+		title: 'Terminkalender',
+		href: '/terminplan',
+		description: 'Übersicht aller wichtigen Termine der Hochschule',
+	},
+	{
+		title: 'Vorlesungen',
+		href: '/vorlesungen',
+		description: 'Übersicht aller Vorlesungen der ersten zwei Semester',
+	},
+	{
+		title: 'Not Found',
+		href: '/not-found',
+		description: 'Übersicht aller wichtigen Termine der Hochschule',
+	},
+];
+
+export default navItems;
